Minor improvement to the way we deal with Mu
I was struggling a bit to understand the precise way that we deal with
recursion. I understood the general approach, but certain things were
quite confusing to me. When writing a function to be used an argument to
`mapNodes`, writing something like

```haskell
foo :: Node -> Node
foo (Node es) = ...
foo n@(Mu _)  = n
..
```

was fine, but we couldn't write

```haskell
foo :: Node -> Node
foo (Node es) = ...
foo n@(Mu _)  = createMu ??
..
```

I understand the reasons, but it still leads to a confusing way to code.
Also, consider this definition of `mapNodes`:

```haskell
    go' :: (Node -> Node) -> Node -> Node
    go' f EmptyNode = EmptyNode
    go' f (Node es) = f $ (Node $ map (\e -> setChildren e $ (map go (edgeChildren e))) es)
    go' f (Mu n)    = f $ (createMu $ const (go n))
    go' f (Rec i)   = case f (Rec i) of
                        Rec _ -> Rec TmpRecNodeId
                        x     -> x
```

The case for `Mu` is really confusing: how come we can simply ignore the
`Node` that `createMu` gives us? (It's only because of how we deal with
`Rec`, but to understand that, you'd need to have a clear understanding
of the internals of what `createMu` does.

Another thing that was a bit confusing is this definition of
`UninternedNode`:

```haskell
data UninternedNode = UninternedNode ![Edge]
                    | UninternedEmptyNode
                    | UninternedMu !Node
                    | UninternedRec
```

Why does it have a `Rec` case? `UninternedNode` is not recursive: we can
never have an `UninternedNode` inside of another one; underneath an
`UninternedMu` we have a `Node`, not an `UninternedNode`. The answer is
of course that this whole thing might be `interned` so live underneath
another `UninteredMu` somewhere above us, but it's still somewhat
unsatifying.

The other thing that was bothering me was that we had these
`TmpRecNodeId` constructors, that exist temporarily, but still were
leading to partial pattern matches everywhere, and hence compiler
warnings (with `-Wall`).

In this commit I've tried to make this a little cleaner.
`UninternedNode` is now defined as

```haskell
data UninternedNode = UninternedNode ![Edge]
                    | UninternedEmptyNode
                    | UninternedMu !(Id -> Node)
```

where the `Rec` case is gone. (Side note: a bang on a list constructor
is a bit of a red flag to me: it only forces the top-most list
constructor, which is typically not what you need. I've left it as it
was for now though.)

`RecNodeId` is now

```haskell
data RecNodeId = RecNodeId { unRecNodeId :: {-# UNPACK #-} !Id }
  deriving ( Eq, Ord, Show, Generic )
```

without its `TmpNodeId` constructor. I then generalized the `Mu` pattern
synonym so that it is bidirectional again, taking care of these `Ids`:

```haskell
pattern Mu :: (Node -> Node) -> Node
pattern Mu f <- (matchMu -> Just f)
  where
    Mu = createMu
```

It's satisfying I think that `Mu` can now use `createMu` in the other
direction. `matchMu` is defined as:

```haskell
matchMu :: Node -> Maybe (Node -> Node)
matchMu (InternedMu oldId n) = Just $ \n' -> substFree oldId n' n
matchMu _otherwise = Nothing

substFree :: Id -> Node -> Node -> Node
substFree old new = go
  where
    go :: Node -> Node
    go = memo (NameTag "substFree") go'
    {-# NOINLINE go #-}

    go' :: Node -> Node
    go' EmptyNode               = EmptyNode
    go' (InternedNode _ es)     = intern $ UninternedNode (map goEdge es)
    go' (InternedMu i n)        = intern $ UninternedMu $ \nid -> go (substFree i (Rec (RecNodeId nid)) n)
    go' n@(Rec (RecNodeId nid)) = if nid == old
                                    then new
                                    else n

    goEdge :: Edge -> Edge
    goEdge e = setChildren e $ map go (edgeChildren e)
```

The traversal in `substFree` is unfortunate, but I don't think we're
doing any work that we weren't doing before; this basically takes the
place of the traversal done by `fillInTmpRecNodes`, which is now gone
entirely.

I think this slight change in API leads to easier to understand code.
For example, function `go'` in `mapNodes` now looks like:

```haskell
    go' :: Node -> Node
    go' EmptyNode = EmptyNode -- TODO: shouldn't this call f too..?
    go' (Node es) = f $ (Node $ map (\e -> setChildren e $ (map go (edgeChildren e))) es)
    go' (Mu n)    = f $ Mu (go . n)
    go' (Rec i)   = f $ Rec i
```

The case for `Mu` and `Rec` are now much easier to understand, I think.
(One question: shouldn't we call `f` in the `EmptyNode` case too...?)

Another example: unfolding was defined as

```haskell
unfoldOuterRec :: Node -> Node
unfoldOuterRec = memo (NameTag "unfoldOuterRec") go
  where
    go :: Node -> Node
    go n@(Mu x) = mapNodes (unrollRec (nodeIdentity n) n) x
    go _        = error "unfoldOuterRec: Must be called on a Mu node"

    unrollRec :: Id -> Node -> Node -> Node
    unrollRec muNodeId muNode n@(Mu _)            = n
    unrollRec muNodeId muNode (Rec (RecNodeId i))
      | i == muNodeId                             = muNode
    unrollRec muNodeId muNode n@(Rec _)             = error $ "unfoldOuterRec: Unexpected Rec node " <> show n
                                                              <> " while unrolling " <> show muNode
    unrollRec _        _      x                   = x
```

I think this is still using the old "single unique Mu" assumption, and
so would need to be updated regardless. However, with the API, it's very
very simple:

```haskell
unfoldOuterRec :: Node -> Node
unfoldOuterRec n@(Mu x) = x n
unfoldOuterRec _        = error "unfoldOuterRec: Must be called on a Mu node"
```

Which is a much more direct definition of unfolding.

One downside is that not all functions can be defined using recursion
using the `Mu` pattern synonym, but I think this can be considered a
feature, not a bug: when we recurse underneath a `Mu`, we have dangling
references. Code that wants to recurse there should explicitly take care
of that. An example is `crush`, which actually just returns `mempty` for
`Rec` nodes, so it doesn't care about dangling references. I've modified
it to use the interned constructors directly.

This commit also makes minimal changes to `toFGL`, enough to update it
for the new API. However, I think `toFGL` still relies on having a
single unique `Mu` node; this commit does not attempt to change that.

The unit tests build and pass with this commit. The benchmarks do _not_
compile, but that is already the case in the `jk-many-mu` branch, so
I've just left them as-is.