Generalize substitution
Instead of substituting a single `Id`, we can now substitute any number
of `RecNodeId`.

This also fixes an interning bug (we were -- intentionally -- not using
the `Node` pattern synonym, and as a result weren't sorting edges before
interning).

This also demonstrates how we can memoize functions such as substitution
independent from an environment. This may not always be advisable
(applying en environment after memoization, if done recursively, is
prohibitively expensive), but for a function like substitution, this
kind of environment application is the whole raison d'etre of the
function. In this case, the different definition does not really have a
different performance as the old (I've verified), but it's an
interesting pattern so I've left it in.